requirements.txt
amqp==5.1.1
APScheduler==3.10.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cron-descriptor==1.2.35
Django==4.2
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
idna==3.4
kombu==5.2.4
prompt-toolkit==3.0.38
psycopg2==2.9.6
PyJWT==2.6.0
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
redis==4.5.4
requests==2.28.2
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.3
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
yarg==0.1.9
